feat: :lipstick: create footer and make Navigation components global
