+++ components/UI/Button/styles.module.css
@@ -9,7 +9,7 @@
   line-height: 24px;
   text-transform: uppercase;
   min-width: 126px;
-  transition: all 200ms ease;
+  transition: all 100ms ease;
   gap: 12px;
 }
