+++ components/global/Footer/styles.module.css
@@ -1,0 +1,3 @@
+.mainFooter {
+  min-height: 300px;
+}
