+++ components/global/Header/styles.module.css
@@ -1,0 +1,3 @@
+.mainHeader {
+  background-color: var(--background-base);
+}
