+++ messages/en.json
@@ -6,6 +6,10 @@
       "blog": "Blog"
     }
   },
+  "Footer": {
+    "sentence": "Be nice :)",
+    "copyright": "Edwin Obando © All rights reserved."
+  },
   "Index": {
     "title": "Home",
     "description": "This is the home page."
